refactor: cohesive state of DifficultyUncleRateEpoch chart
